拆分store (#15)
* 创建userStore
